magi join: narrow JoinSlot + dense build-payload arena; XXXLARGE join tier
The join hash slot shrinks from a fat AggSlot64 to {key, index,
payload_idx} (16B for uint64 keys, 32B for u128): BuildInsert claims a
dense payload row via an atomicAdd bump counter and stores its index in
the slot; ProbeEmit reads payloads through that index. Halving the slot
lets the always-resident join arena hold a 128M-slot XXXLARGE tier in
the same 4GB the 64M fat-slot tier used — needed for build sides like
SF100 Q9's orders (150M rows, ~75M keys/GPU), which hung the old table
(unbounded linear probe on a full 64M table). Probing is now bounded at
4096 steps and a pre-launch capacity check throws (DuckDB fallback)
when the per-GPU estimate exceeds the largest tier.

magi_distributed_groupby.hpp gains N_SLOTS_XXXLARGE / TableSize::XXXLARGE
(shared with the groupby ladder) and the GroupbyXxxlEnabled() switch the
groupby-side commit wires up.

Also: channel-cost print uses a signed difference (free can grow across
init when another process exits).

Verified: SF100 Q9 158ms warm (was hang); custom join+groupby query
row-identical to CPU at SF50; full SF50/SF100 protocol sweeps show zero
regression.

+++ src/legacy/cuda/magi/magi_runtime.cu
@@ -279,7 +279,9 @@ void MagiInitOnce()
     size_t f = 0, t = 0; cudaMemGetInfo(&f, &t);
     std::printf("[magi] GPU %d channel cost: %.2f GB (free %.2f -> %.2f GB), "
                 "user_grid=%zu intra=%zuMiB inter=%zuMiB\n",
-                s.gpu_ids[i], (free_before[i] - f) / (1024.0 * 1024 * 1024),
+                s.gpu_ids[i],
+                (static_cast<double>(free_before[i]) - static_cast<double>(f)) /
+                    (1024.0 * 1024 * 1024),
                 free_before[i] / (1024.0 * 1024 * 1024), f / (1024.0 * 1024 * 1024),
                 USER_KERNEL_GRID_SIZE,
                 KBUFFERING_INTRA_PARTITION_SIZE >> 20,
